processing_engine/requirements.txt
dspy
supabase
pytest
python-dotenv
ddgs
